works

+++ base.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <vector>
 
-template <int height, int width>
+template <size_t height, size_t width>
 struct base {
     static const int h = height;
     static const int w = width;
